Daka-js is now a very fancy and probably buggy calculator
This is a big commit. I had to get a lot of things in place before I could see if
things actually worked, but now we have the skeleton in place. We can parse and run
basic arithmetic expression and have support for return statements. It's enough that
we can start testing the language and seeing what works.